Fix compiler warnings and the usage of read(2)
When current semu is compiled, compiler shows following warnings about
ignored return values:

  semu.c: In function ‘uart_thread_func’:
  semu.c:308:9: warning: ignoring return value of ‘read’ declared...
    308 |         read(STDIN_FILENO, &c, 1);
        |         ^~~~~~~~~~~~~~~~~~~~~~~~~
  semu.c: In function ‘read_file’:
  semu.c:1306:5: warning: ignoring return value of ‘fread’ declared...
   1306 |     fread(content, fsize, 1, f);
        |     ^~~~~~~~~~~~~~~~~~~~~~~~~~~

Besides, the first read() API is used to simulate an external UART RX
signal. If there is no data when semu calls read(), it is uncessary to
execute the following operations which simulates an UART interrupt.

This patch including:
1. Check the return value when semu calls read API to read content from
STDIN. If read() returns an error or EOF, it means there is no external
input (UART signal) now. Therefore we just continue the loop execution.

2. Check the return value when semu calls fread() API to read content
from binary or image. If the size of read data is less than requested
size, a fatal error is triggered to exit semu.