Refine riscv-tests binary generations and tracking
This patch decouples the generation of riscv-tests specific files and
the target for running ISA tests. It is allowed to build riscv-tests
files in advance (likely on Linux machines) and trigger ISA tests in
different environments.

The tracking of corresponding raw binary files is improved, and it
should take less time now.