Check address alignment of AMO instructions
For AMOs, the A extension requires that the address held in rs1 be
naturally aligned to the size of the operand (i.e., eight-byte aligned
for 64-bit words and four-byte aligned for 32-bit words). If the address
is not naturally aligned, an address-misaligned exception or an
access-fault exception will be generated.