Bump buildroot and binutils version to stable tag
The stable tag of buildroot is 2024.02.x currently. Thus, upgrade the
default buildroot version of 2023.11.x to 2024.02.x .

The commit 1391c99 in branch 2024.02.x of buildroot dropped the support
of 2.38.x series of binutils, set the default version to 2.40.x and
addded support of version 2.41.x. To ensure the compliance, upgrade the
version of binutils in the default configuration of buildroot.

This commit also fix the bumping to the wrong version of binutils in
commit 2c1e29c.