Preliminary allow PLIC to support 32 sources to 32 contexts
By changing `ie` from `uint32_t` to an array of `uint32_t`, this can
simulate different interrupt enable registers for different contexts.
For example, `ie[0]` simulates the interrupt enable register for the
first context, which is at 0x002000 in the address map, and `ie[1]`
simulates the second context with the address map at 0x002080.