Preliminary ACLINT implementation
In this commit, based on the implementation of CLINT, an preliminary
ACLINT is implemented, including the basic logic for operating `mtimer`,
`mswi`, and `sswi`.

Additionally, introduces a new script `gen-hart-dts.py` for generating
device-tree configurations and renames the original script to
`gen-clint-dts.py`. The new script is currently unused.

A new macro SEMU_FEATURE_ACLINT is also added to feature.h, allowing
the replacement of CLINT with ACLINT in the implementation.

Currently, due to the lack of implementation, the introduced ACLINT
uses only supervisor-level IPI. Therefore, although the logic for mswi
is implemented, it is not being used at the moment.

You can test it directly using the command `make check SMP=n`, where n
is the number of harts you want to simulate.