Bump Buildroot to the 2024.11.1 release
The GCC 14 now enforces a stricter error level by treating the -Wimplicit-int
warning as an error by default. This change causes a compilation issue with the
libsha1 library, which is necessary for enabling the X.Org Server.

Upgrading to Buildroot version 2024.11.1 resolves this issue, as it includes a
patch [1] that addresses the problem.

[1] https://github.com/buildroot/buildroot/blob/65137580ca9ed060c59544b9859fc64
8e902c694/package/libsha1/0001-test-fix-gcc-14.x-compile-implicit-int.patch