Enable user-mode networking through SLIRP
This commit integrates 'minislirp' as submodule and enables support
for the new 'user' network device, which uses SLIRP for user-mode
networking.

An internal pipe mechanism is introduced to monitor incoming data.
The main loop polls the pipe for availabe data, and any incoming
data is forwarded to the virtio-net device for processing.

Additionally, this commit introduces special address used by SLIRP
for network configuration [1]:
- **10.0.2.0**: The SLIRP "on-line" configuration address.
- **10.0.2.2**: An alias for the host running SLIRP.
- **10.0.2.3**: An alias for the DNS address.
- **10.0.2.15**: A recommended address for the PC running SLIRP.

[1]: https://github.com/kost/slirp/blob/master/docs/slirp.doc