Support mounting rootfs from virtio-blk via initrd
Due to the 1 GiB kernel space limit on RV32-Linux, embedding large root
filesystems in the kernel image can be impractical for certain testing
scenarios.

This commit introduces a script to generate a mountable ext4 rootfs image and
modifies the init script to support switching the root filesystem to a
virtio-blk device.

During boot, the init script checks whether /dev/vda contains a valid rootfs.
If so, it switches the root to the virtio-blk device. Otherwise, it falls back
to the default initramfs.